fix Logreg add results
